requirements.txt
dj-extensions==0.1.6
Django==1.11.15
django-appconf==1.0.2
django-ckeditor==5.4.0
django-easy-select2==1.3.4
django-extensions==1.9.9
django-js-asset==1.0.0
django-stdimage==3.0.1
djangorestframework==3.8.2
mysqlclient==1.3.12
olefile==0.44
pilkit==2.0
Pillow==4.3.0
progressbar2==3.35.2
pydot==1.2.4
pygraphviz==1.3.1
pyparsing==2.2.0
python-utils==2.3.0
pytz==2017.2
six==1.11.0
typing==3.6.4
